Add getGamePlayer
took me forever to find out why it didnt work. It didnt work because i used an return in a foreach. Apparently it doesnt stop so it continues and return the values of the last item. Use find instead it stops at first returned value. Lots of time wasted but atleast i learned something.